Use int as age

+++ src/test/resources/database/oracle/CREATE_TEST_TABLE.sql
@@ -2,6 +2,6 @@ CREATE TABLE STUDENT
 (
     id      varchar(255),
     name    varchar(255),
-    age     number(3),
+    age     int,
     average number
 )
